backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
PyGithub
redis
python-multipart
